backend/requirements.txt
torch
torchvision
tqdm
scikit-learn
Pillow
fastapi
uvicorn[standard]
numpy
python-multipart